requirements.txt
# -----------------------------
# Core Backend
# -----------------------------
fastapi==0.110.0
uvicorn==0.29.0
python-multipart==0.0.9
python-dotenv==1.0.1
requests==2.31.0

# -----------------------------
# Vision / OCR
# -----------------------------
pillow==10.1.0
pymupdf==1.23.8

# -----------------------------
# LlamaIndex
# -----------------------------
llama-index==0.12.4
llama-cloud-services==0.6.91
# -----------
# OpenAI SDK
# -----------------------------
openai==1.59.7

# -----------------------------
# Transformers (CPU)
# -----------------------------
transformers==4.36.2

# -----------------------------
# Frontend
# -----------------------------
streamlit==1.31.1
